BSC Subgraph Improvement
