--- BSC/controller/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "liquid-original-bsc-testnet",
-  "license": "UNLICENSED",
-  "scripts": {
-    "codegen": "graph codegen",
-    "build": "graph build",
-    "deploy": "graph deploy --node https://api.thegraph.com/deploy/ shreyas3336/liquid-original-bsc-testnet",
-    "create-local": "graph create --node http://localhost:8020/ liquid-original-bsc-testnet",
-    "remove-local": "graph remove --node http://localhost:8020/ liquid-original-bsc-testnet",
-    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 liquid-original-bsc-testnet"
-  },
-  "dependencies": {
-    "@graphprotocol/graph-cli": "0.23.2",
-    "@graphprotocol/graph-ts": "0.23.1"
-  }
-}
