--- BSC/master/schema.graphql
@@ -1,9 +1,0 @@
-type Project @entity {
-  id: ID!
-  projectOwnerAddress: Bytes! 
-  projectName: String!
-  projectTokenAddress: Bytes!
-  projectTokenTicker: String!
-  projectDocHash: String!
-  projectTokenDecimal: BigInt!
-}
