+++ BSC/subgraph.yaml
@@ -1,0 +1,82 @@
+specVersion: 0.0.4
+schema:
+  file: ./schema.graphql
+dataSources:
+  - kind: ethereum
+    name: Master
+    network: chapel
+    source:
+      address: "0x7f8EDc9aC14BD669A5f68eC226c7c98A66BCc076"
+      abi: Master
+      startBlock: 15219520
+    mapping:
+      kind: ethereum/events
+      apiVersion: 0.0.5
+      language: wasm/assemblyscript
+      entities:
+        - ProjectInfo
+      abis:
+        - name: Master
+          file: ./abis/Master.json
+      eventHandlers:
+        - event: ProjectInfo(string,indexed address,string,string,address,uint256)
+          handler: handleProjectInfoMaster
+      file: ./src/master.ts
+  - kind: ethereum
+    name: Controller
+    network: chapel
+    source:
+      address: "0x765D6a9fd7889F5D7414311799dAFf5E0eC34099"
+      abi: Controller
+      startBlock: 15219527
+    mapping:
+      kind: ethereum/events
+      apiVersion: 0.0.5
+      language: wasm/assemblyscript
+      entities:
+        - CreateVest
+        - TransferWrapped
+        - Withdraw
+        - ProjectInfo
+      abis:
+        - name: Controller
+          file: ./abis/Controller.json
+      eventHandlers:
+        - event: CreateVest(indexed address,address,address,uint256,uint256,address,string,bool)
+          handler: handleCreateVestController
+        - event: TransferWrapped(address,indexed address,address,uint256)
+          handler: handleTransferWrappedController
+        - event: Withdraw(indexed address,uint256,address)
+          handler: handleWithdrawController
+        - event: ProjectInfo(indexed address,string,address,uint256)
+          handler: handleProjectInfoController
+      file: ./src/controller.ts
+  - kind: ethereum/contract
+    name: Vesting
+    network: chapel
+    source:
+      address: "0x1f2aEf238D61Dc7b16A4aD78b54d70F262e14483"
+      abi: Vesting
+      startBlock: 15219549
+    mapping:
+      kind: ethereum/events
+      apiVersion: 0.0.5
+      language: wasm/assemblyscript
+      entities:
+        - CreateVest
+        - TransferLock
+        - Withdraw
+        - ProjectInfo
+      abis:
+        - name: Vesting
+          file: ./abis/Vesting.json
+      eventHandlers:
+        - event: CreateVest(uint256,indexed address,address,address,uint256,uint256)
+          handler: handleCreateVestLock
+        - event: TransferLock(uint256,address,indexed address,address,uint256,uint256)
+          handler: handleTransferLock
+        - event: Withdraw(uint256,indexed address,uint256,address,uint256)
+          handler: handleWithdrawLock
+        - event: ProjectInfo(indexed address,indexed address)
+          handler: handleProjectInfoVesting
+      file: ./src/vesting.ts
