--- BSC/vesting/subgraph.yaml
@@ -1,45 +1,0 @@
-specVersion: 0.0.2
-schema:
-  file: ./schema.graphql
-dataSources:
-  - kind: ethereum/contract
-    name: Vesting
-    network: chapel
-    source:
-      address: "0x1f2aEf238D61Dc7b16A4aD78b54d70F262e14483"
-      abi: Vesting
-      startBlock: 15219549
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.5
-      language: wasm/assemblyscript
-      entities:
-        - AdminChanged
-        - BeaconUpgraded
-        - CreateVest
-        - OwnershipTransferred
-        - ProjectInfo
-        - TransferLock
-        - Upgraded
-        - Withdraw
-      abis:
-        - name: Vesting
-          file: ./abis/Vesting.json
-      eventHandlers:
-        - event: AdminChanged(address,address)
-          handler: handleAdminChanged
-        - event: BeaconUpgraded(indexed address)
-          handler: handleBeaconUpgraded
-        - event: CreateVest(uint256,indexed address,address,address,uint256,uint256)
-          handler: handleCreateVest
-        - event: OwnershipTransferred(indexed address,indexed address)
-          handler: handleOwnershipTransferred
-        - event: ProjectInfo(indexed address,indexed address)
-          handler: handleProjectInfo
-        - event: TransferLock(address,indexed address,address,uint256,uint256)
-          handler: handleTransferLock
-        - event: Upgraded(indexed address)
-          handler: handleUpgraded
-        - event: Withdraw(indexed address,uint256,address,uint256)
-          handler: handleWithdraw
-      file: ./src/mapping.ts
